+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Improve map load time by requesting the icon set earlier and lazy parsing icons on-demand.
 - Add support for the globe to mercator transition when using the global placement.
 - Add support for symbol fading when using the global placement.
+- Add debug visualization support when using the global placement.
 
 ### Bug fixes 🐞